UpdateUser
Update user

+++ Reto5/Reto5 - SQL/CreateUsers.sql
@@ -8,7 +8,7 @@ CREATE TABLE IF NOT EXISTS Users (
     user_name VARCHAR(60) NULL,
     user_lastname VARCHAR(60) NULL,
     user_email VARCHAR(60) NULL DEFAULT "user@123.com",
-    user_phone INT NULL,
+    user_phone LONG NULL,
     user_birthday DATE NULL DEFAULT "1900-01-01",
     PRIMARY KEY (user_username)
 );
